Crate orml_tokens

Tokens Module

Overview

The tokens module provides fungible multi-currency functionality that implements MultiCurrency trait.

The tokens module provides functions for:

Querying and setting the balance of a given account.

Getting and managing total issuance.

Balance transfer between accounts.

Depositing and withdrawing balance.

Slashing an account balance.

Implementations

The tokens module provides implementations for following traits.

MultiCurrency - Abstraction over a fungible multi-currency system.

MultiCurrencyExtended - Extended MultiCurrency with additional helper types and methods, like updating balance by a given signed integer amount.

Interface

Dispatchable Functions

transfer - Transfer some balance to another account.

transfer_all - Transfer all balance to another account.

Genesis Config

The tokens module depends on the GenesisConfig. Endowed accounts could be configured in genesis configs.

Re-exports

pub use module::*;

Modules

module

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

AccountData

balance information for an account.

BalanceLock

A single lock on a balance. There can be many of these on an account and they “overlap”, so the same balance is frozen by multiple locks.

BurnDust

Combiner

CurrencyAdapter

Mapper

NegativeImbalance

Opaque, move-only struct with private fields that serves as a token denoting that funds have been destroyed without any equal and opposite accounting.

PositiveImbalance

Opaque, move-only struct with private fields that serves as a token denoting that funds have been created without any equal and opposite accounting.

TransferDust

Traits

ConvertBalance

WeightInfo

Weight functions needed for orml_tokens.