Struct orml_tokens::module::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn transfer(
origin: OriginFor<T>,
dest: <T::Lookup as StaticLookup>::Source,
currency_id: T::CurrencyId,
amount: T::Balance
) -> DispatchResult

Transfer some liquid free balance to another account.

transfer will set the FreeBalance of the sender and receiver. It will decrease the total issuance of the system by the TransferFee. If the sender’s account is below the existential deposit as a result of the transfer, the account will be reaped.

The dispatch origin for this call must be Signed by the transactor.

dest: The recipient of the transfer.

currency_id: currency type.

amount: free balance amount to tranfer.

pub fn transfer_all(
origin: OriginFor<T>,
dest: <T::Lookup as StaticLookup>::Source,
currency_id: T::CurrencyId,
keep_alive: bool
) -> DispatchResult

Transfer all remaining balance to the given account.

NOTE: This function only attempts to transfer transferable balances. This means that any locked, reserved, or existential deposits (when keep_alive is true), will not be transferred by this function. To ensure that this function results in a killed account, you might need to prepare the account by removing any reference counters, storage deposits, etc…

The dispatch origin for this call must be Signed by the transactor.

dest: The recipient of the transfer.

currency_id: currency type.

keep_alive: A boolean to determine if the transfer_all operation should send all of the funds the account has, causing the sender account to be killed (false), or transfer everything except at least the existential deposit, which will guarantee to keep the sender account alive (true).

pub fn transfer_keep_alive(
origin: OriginFor<T>,
dest: <T::Lookup as StaticLookup>::Source,
currency_id: T::CurrencyId,
amount: T::Balance
) -> DispatchResultWithPostInfo

Same as the [transfer] call, but with a check that the transfer will not kill the origin account.

99% of the time you want [transfer] instead.

The dispatch origin for this call must be Signed by the transactor.

dest: The recipient of the transfer.

currency_id: currency type.

amount: free balance amount to tranfer.

pub fn force_transfer(
origin: OriginFor<T>,
source: <T::Lookup as StaticLookup>::Source,
dest: <T::Lookup as StaticLookup>::Source,
currency_id: T::CurrencyId,
amount: T::Balance
) -> DispatchResult

Exactly as transfer, except the origin must be root and the source account may be specified.

The dispatch origin for this call must be Root.

source: The sender of the transfer.

dest: The recipient of the transfer.

currency_id: currency type.

amount: free balance amount to tranfer.

pub fn set_balance(
origin: OriginFor<T>,
who: <T::Lookup as StaticLookup>::Source,
currency_id: T::CurrencyId,
new_free: T::Balance,
new_reserved: T::Balance
) -> DispatchResult

Set the balances of a given account.

This will alter FreeBalance and ReservedBalance in storage. it will also decrease the total issuance of the system (TotalIssuance). If the new free or reserved balance is below the existential deposit, it will reap the AccountInfo.

The dispatch origin for this call is root.

impl<T: Config> Pallet<T>

pub fn total_issuance<KArg>(k: KArg) -> T::Balancewhere
KArg: EncodeLike<T::CurrencyId>,

The total issuance of a token type.

impl<T: Config> Pallet<T>

pub fn locks<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> BoundedVec<BalanceLock<T::Balance>, T::MaxLocks>where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<T::CurrencyId>,

Any liquidity locks of a token type under an account. NOTE: Should only be accessed when setting, changing and freeing a lock.

impl<T: Config> Pallet<T>

pub fn accounts<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> AccountData<T::Balance>where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<T::CurrencyId>,

The balance of a token type under an account.

NOTE: If the total is ever zero, decrease account ref account.

NOTE: This is only used in the case that this module is used to store balances.

impl<T: Config> Pallet<T>

pub fn reserves<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> BoundedVec<ReserveData<T::ReserveIdentifier, T::Balance>, T::MaxReserves>where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<T::CurrencyId>,

Named reserves on some account balances.

impl<T: Config> Pallet<T>

pub fn map_tx_count<KArg>(k: KArg) -> Option<u64>where
KArg: EncodeLike<T::AccountId>,