Trait orml_tokens::module::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen;
    type Amount: Signed + TryInto<Self::Balance> + TryFrom<Self::Balance> + Parameter + Member + SimpleArithmetic + Default + Copy + MaybeSerializeDeserialize;
    type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo;
    type WeightInfo: WeightInfo;
    type ExistentialDeposits: GetByKey<Self::CurrencyId, Self::Balance>;
    type OnDust: OnDust<Self::AccountId, Self::CurrencyId, Self::Balance>;
    type MaxLocks: Get<u32>;
    type DustRemovalWhitelist: Contains<Self::AccountId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen

The balance type

type Amount: Signed + TryInto<Self::Balance> + TryFrom<Self::Balance> + Parameter + Member + SimpleArithmetic + Default + Copy + MaybeSerializeDeserialize

The amount type, should be signed version of Balance

type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo

The currency ID type

type WeightInfo: WeightInfo

Weight information for extrinsics in this module.

type ExistentialDeposits: GetByKey<Self::CurrencyId, Self::Balance>

The minimum amount required to keep an account. It’s deprecated to config 0 as ED for any currency_id, zero ED will retain account even if its total is zero. Since accounts of orml_tokens are also used as providers of System::AccountInfo, zero ED may cause some problems.

type OnDust: OnDust<Self::AccountId, Self::CurrencyId, Self::Balance>

Handler to burn or transfer account’s dust

type MaxLocks: Get<u32>

type DustRemovalWhitelist: Contains<Self::AccountId>